Move fixupUserHomeProperty() to utils
To be closer to the related getUserHomeDirectory() function, and for
potential reuse e.g. in helper-cli.
