+++ utils/src/main/kotlin/Utils.kt
@@ -187,6 +187,24 @@ fun getPathFromEnvironment(executable: String): File? {
 }
 
 /**
+ * Check if the "user.home" property is set to a sane value and otherwise set it to the value of an (OS-specific)
+ * environment variable for the user home directory. This works around the issue that esp. in certain Docker scenarios
+ * "user.home" is set to "?", see https://bugs.openjdk.java.net/browse/JDK-8193433 for some background information.
+ */
+fun fixupUserHomeProperty() {
+    val userHome = System.getProperty("user.home")
+    val checkedUserHome = sequenceOf(
+        userHome,
+        Os.env["HOME"],
+        Os.env["USERPROFILE"]
+    ).find {
+        !it.isNullOrBlank() && it != "?"
+    }
+
+    if (checkedUserHome != null && checkedUserHome != userHome) System.setProperty("user.home", checkedUserHome)
+}
+
+/**
  * Return the current user's home directory.
  */
 fun getUserHomeDirectory() = File(System.getProperty("user.home"))
